Allow passing NodeStacks to nested visitors (#285)
This PR adds an `options` objects to most core visitors (or refactors optional arguments into an `options` objects).

Within this `options` object, it offers the ability for any visitor that uses a `NodeStack` to use the provided one instead of a brand new one. This enables visitors to call others visitors during their traversal whilst all sharing the same `NodeStack` and not losing track of where we are in the tree.