Include CLI in main Codama library (#777)
This PR adds the `codama` CLI back to the main library so it is available when consumers install `codama` — instead of having to additionally install `@codama/cli`.

This change was reverted before (See <https://github.com/codama-idl/codama/pull/596>) because `@codama/cli` contained various dependencies (such as renderers) that would then be linked with the main library when versioning package. This caused issues such as: when bumping a renderer package, the node package would need to also be bumped. This limitation is now gone since these additionally dependencies have been removed throughout the PR stack that this PR is part of.